handling database mismatches

+++ projects/capstone/vaccines_monitor/patients.py
@@ -39,7 +39,10 @@ def deletePatient(id):
     if not patient:
         abort(404)
     try:
-        patient_user_management.deletePatientUser(patient.user_id)
+        if not patient_user_management.getPatientUser(patient.user_id):
+            print("Note: patient doesnt exist in Auth0 database")
+        else:
+            patient_user_management.deletePatientUser(patient.user_id)
         patient.delete()
         return jsonify({
         'success': True
